Move src/components/elementsList to src/apps/mydb/elements/list/
