+++ app/packs/src/apps/mydb/elements/list/ElementAllCheckbox.js
